+++ faizdb-server/src/wire/mod.rs
@@ -8,5 +8,5 @@ pub mod op_query;
 
 pub mod postgres;
 
-pub use listener::run_wire_server;
-pub use postgres::run_postgres_server;
+pub use listener::{run_wire_server, run_wire_server_with_shutdown};
+pub use postgres::{run_postgres_server, run_postgres_server_with_shutdown};